UP: User routes moved
